requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2

# Document parsing
python-docx==0.8.11
openpyxl==3.1.2
PyPDF2==3.0.1
python-pptx==0.6.21

# AI / ML
google-generativeai==0.3.0
sentence-transformers==2.2.2
numpy==1.24.3

# Database
sqlite3

# Environment
python-dotenv==1.0.0

# Utilities
requests==2.31.0
